Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0098-validate-binary-search-tree/0098-validate-binary-search-tree.cpp
@@ -18,6 +18,6 @@ public:
     }
     bool isValidBST(TreeNode* root) {
         if(!root)return true;
-        return isValid(root,INT_MIN,INT_MAX);
+        return isValid(root,LONG_MIN,LONG_MAX);
     }
 };
